Add benchmarking and profiling

+++ sources/record.go
@@ -4,7 +4,7 @@ import "encoding/json"
 
 type Record struct {
 	Uid string `json:"uid"`
-	Timestamp float64 `json:"ts"`
+	Timestamp uint64 `json:"ts"`
 }
 
 func ParseRecord(raw []byte) (r Record, err error) {
